requirements.txt
streamlit
boto3
python-dotenv
pandas
matplotlib
seaborn
